MATH_final/requirements.txt
# Core
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.20.0
datasets>=2.14.0
numpy>=1.21.0
psutil>=5.9.0
matplotlib>=3.7.0
vllm>=0.4.0
requests>=2.31.0

# LlamaIndex
llama-index-core>=0.10.0
llama-index-llms-huggingface>=0.1.0

# LangChain
langchain>=0.1.0
langchain-community>=0.0.10

# Adaptive context strategy (Math-BERT embeddings)
sentence-transformers>=2.2.0
